feat: migrate CreateSellOrderForm to react-hook-form (#2132)
* feat: migrate CreateSellOrderForm to react-hook-form

* chore: remove old files and migrate story

* fix: typing issue

* chore: clean unused contants file
