+++ web-marketplace/src/pages/Dashboard/MyEcocredits/MyEcocredits.tsx
@@ -19,7 +19,6 @@ import {
   BASKET_TAKE_TITLE,
   BasketTakeModal,
 } from 'web-components/lib/components/modal/BasketTakeModal';
-import { CreateSellOrderModal } from 'web-components/lib/components/modal/CreateSellOrderModal';
 import { ProcessingModal } from 'web-components/lib/components/modal/ProcessingModal';
 import { TxErrorModal } from 'web-components/lib/components/modal/TxErrorModal';
 import { Item } from 'web-components/lib/components/modal/TxModal';
@@ -40,6 +39,7 @@ import { chainInfo } from 'lib/wallet/chainInfo/chainInfo';
 
 import { Link } from 'components/atoms';
 import WithLoader from 'components/atoms/WithLoader';
+import { CreateSellOrderModal } from 'components/organisms/CreateSellOrderModal/CreateSellOrderModal';
 import { CreditRetireModal } from 'components/organisms/Modals/CreditRetireModal/CreditRetireModal';
 import { CREDIT_RETIRE_TITLE } from 'components/organisms/Modals/CreditRetireModal/CreditRetireModal.constants';
 import {
